mod - Club -> Institution으로 API 통합 - 검색 UX 수정
