+++ src/api/types/institution.ts
@@ -1,3 +1,6 @@
+import { CategoryType } from '@/components/club/constants'
+import { ClubInterface } from '@/types/club'
+
 export interface ClubProfileProps {
   name: string
   summary: string
@@ -5,3 +8,20 @@ export interface ClubProfileProps {
   category: string
   ranking: number
 }
+
+export interface GetClubRequest {
+  sortBy: 'like' | null
+  wishList: boolean
+  category: CategoryType
+  keyword: string | null
+  isLogin: boolean
+}
+
+export type GetClubResponse = ClubInterface[]
+
+export interface PostClubLikeRequest {
+  queryParams: GetClubRequest
+  clubId: number
+}
+
+export type PostClubLikeResponse = ClubInterface
